requirements.txt
# Requirements for Amazon Buy Box Analyzer
# Install with: pip install -r requirements.txt

# Core Data Processing
pandas>=1.5.0                    # Data manipulation and analysis
numpy>=1.21.0                    # Numerical computing

# Amazon SP-API
python-amazon-sp-api>=1.0.0      # Amazon Selling Partner API wrapper

# API and HTTP Operations
requests>=2.28.0                 # HTTP library for API calls
urllib3>=1.26.0                  # HTTP client
tenacity>=8.0.0                  # Retry with exponential backoff

# Environment and Configuration
python-dotenv>=0.19.0            # Environment variable management

# File Processing
openpyxl>=3.0.10                 # Excel file operations
xlsxwriter>=3.0.3                # Excel file writing with formatting

# Testing Framework
pytest>=7.0.0                    # Testing framework
pytest-cov>=4.0.0                # Coverage reporting
pytest-mock>=3.10.0              # Mocking utilities

# Development Tools
black>=22.0.0                    # Code formatting
flake8>=5.0.0                    # Code linting
mypy>=0.991                      # Type checking